Migrate embedded Namecoin Core to Namecoin Legacy
Sadly Namecoin Core does not support queueing of nameupdates. This would
require queueing on our side and most probably will be implemented in
future release, this is just a temporary fix.

This commit adds default rpcpassword as well in initial namecoind
configuration. We allow only 127.0.0.1 connections, but this should be
really generated randomly. As this is a test/prerelease, we are fine
with this, and will just verify if rpcpassword is unchanged in further
releases and notify user accordingly.